tests: Break up test_ospf_flood_reduction.py script
This test run takes 7.2 minutes on my machine.  The
entirety of all tests takes just under 10 minutes now.
Let's break up this test into smaller chunks so that
the test system can take better advantage of parallelization.
